voice-agent/pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[google,openai]~=1.4",
    "livekit-plugins-deepgram>=1.4.5",
    "livekit-plugins-noise-cancellation>=0.2.5",
    "livekit-plugins-silero>=1.4.5",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
